Finish organizing classes with naming conventions and prep for update sql
