+++ scripts/plant_def_db_from_spreadsheet.py
@@ -11,8 +11,8 @@
 # pylint: disable=missing-function-docstring
 import csv
 import os
-from common_classes.paragraphs_to_db import ParagraphsToDatabase
-from common_classes.paragraph_db_input_creator import ParagraphDbInputCreator
+from common_classes.para_db_create_process import ParaDbCreateProcess
+from common_classes.para_db_update_process import ParaDbUpdateProcess
 import helpers.no_import_common_class.paragraph_helpers as no_import_helper
 from portfolio.settings import BASE_DIR
 
@@ -31,10 +31,10 @@ def run(*args):
         > python manage.py runscript -v3 plant_def_db_from_spreadsheet --script-args create_para='yes'
     '''
     print(f'csv input = {IN_CSV_FILE}')
-    json_creator = ParagraphDbInputCreator(title='Botany Definitions')
+    json_creator = ParaDbUpdateProcess(title='Botany Definitions')
     json_creator = process_csv(json_creator)
     if create_para_directly(args):
-        para_to_db = ParagraphsToDatabase()
+        para_to_db = ParaDbCreateProcess()
         para_to_db.dictionary_to_db(json_creator.output)
     else:
         json_creator.write_json_file()
@@ -57,9 +57,9 @@ def process_row(row, json_creator, line_count):
     link_text = form_link_text(row['link_text'], row['add_text'], row['term'])
     url = form_url(row['link'], row['add_text'], row['term'])
     text = [row['desc']]
-    ref = ParagraphDbInputCreator.reference_dictionary(link_text, url)
-    para = ParagraphDbInputCreator.paragraph_dictionary(para_id, text, subtitle)
-    ref_link_para = ParagraphDbInputCreator.ref_link_para_dictionary(para_id, link_text)
+    ref = ParaDbUpdateProcess.reference_dictionary(link_text, url)
+    para = ParaDbUpdateProcess.paragraph_dictionary(para_id, text, subtitle)
+    ref_link_para = ParaDbUpdateProcess.ref_link_para_dictionary(para_id, link_text)
     json_creator.assign_output(ref, para, ref_link_para)
     return json_creator
